move fs create readable stream to ipc-functions
