+++ packages/coinstac-ui/app/main/utils/ipc-functions.js
@@ -1,4 +1,62 @@
-const electron = require('electron');
+
+const fs = require('fs');
+const CsvReadableStream = require('csv-reader');
+const {
+  ipcMain,
+  Notification,
+  quit,
+  relaunch,
+} = require('electron');
+
+ipcMain.handle('quit', () => quit());
+ipcMain.handle('relaunch', () => {
+  relaunch();
+  quit();
+});
+ipcMain.handle('parseCsv', (event, files) => {
+  const readPromises = files.map(file => new Promise((resolve, reject) => {
+    try {
+      const inputStream = fs.createReadStream(file, 'utf8');
+
+      const data = {};
+      let header = [];
+
+      inputStream
+        .pipe(new CsvReadableStream({
+          parseNumbers: true, parseBooleans: true, trim: true, skipHeader: true,
+        }))
+        .on('header', (headerRow) => {
+          header = headerRow;
+        })
+        .on('data', (row) => {
+          const rowData = {};
+
+          row.forEach((cell, i) => {
+            if (i === 0) {
+              return;
+            }
+
+            rowData[header[i]] = cell;
+          });
+
+          // First column of each row serves as the row id
+          data[row[0]] = rowData;
+        })
+        .on('end', () => {
+          // Remove the first column from the header as it's just an id and does not need to
+          // be shown to the user
+          header = header.filter((col, i) => i !== 0);
+
+          resolve({ header, data });
+        });
+    } catch (error) {
+      reject(error);
+    }
+  }));
+
+  return Promise.all(readPromises);
+});
+
 
 module.exports = {
   manualDirectorySelection(path) {
@@ -8,7 +66,7 @@ module.exports = {
     return core.constructor.getJSONSchema(filePath[0]);
   },
   sendNotification(title, body) {
-    const notification = new electron.Notification({ title, body });
+    const notification = new Notification({ title, body });
     notification.show();
   },
 };
